Move Bluetooth LE module to "symmetric" design
This leverages a gatt server AND client on both ends of the connection.
Notifications are used for packets sent both directions.

Add a transaction system for specifying interactions between devices as
a finite state machine.

This includes multiple hacky fixes for gatt transactions

- multiple fixes for circular buffer
- handle completed streams in InputStreamCallback
- simple manual notification test
- request more from InputStreamFlowableSubscriber subscription on read()
- change subscription order when parsing packet
- don't delay server notifications (oops)
- remove most subscribeOn scheduler
- cleanup, better logging, fix potential race condition with setting stage in LeDeviceSession
- do not attempt handshake twice
- log when disposing gatt server
- do not set default reply
- use RxAndroidBle's QUICK_SETUP flag for notifications
- rewrite gatt server reverse connection

Fix multiple timing/subscription issues with protobuf transaction workflow
    - extremely hacky fix for subscription gap issue
    - set global error handler
    - cleanup and restructure

- automated cleanup

NOTE:
Notification inner observable is a hot observable, with the first values
emitted VERY early, which means that the values are emitted before any reasonable
observer can subscribe. A hacky workaround is to convert to flowable with "buffer"
backpressure strategy and subscribe to that flowable instead